requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
click==8.2.1
colorama==0.4.6
databases==0.9.0
fastapi==0.115.12
greenlet==3.2.3
h11==0.16.0
idna==3.10
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-decouple==3.8
python-dotenv==1.1.0
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
